Working on porting point assessments from an integer that represents a float to the BigDecimal class BigDecimal is designed for such use cases

+++ src/main/kotlin/com/iprp/backend/controller/RestStudent.java
@@ -137,7 +137,7 @@ public class RestStudent {
         }
 
         return dm.updateReview(
-            authentication.getName(), id, (String) json.get("feedback"), (List<Integer>) json.get("points")
+            authentication.getName(), id, (String) json.get("feedback"), (List<Double>) json.get("points")
         );
     }
 }
